Minor sidemenu style changes
